refactor: Modulo user refatorizado completamente

+++ src/useCases/User/DeleteUserUseCase/DeleteUserUseCase.ts
@@ -1,20 +1,24 @@
 import { IUserRepository } from "@/repositories/User/IUserRepository";
 import { DeleteUserDTO } from "./DeleteUserDTO";
 import { ErrorUserDoesNotExist } from "@/erros/ErrorUserDoesNotExist";
+import { ErrorPermission } from "@/erros/User/ErrorPermission";
 
-export class DeleteUserUseCase{
-    constructor(private userRepository: IUserRepository){}
+export class DeleteUserUseCase {
+  constructor(private userRepository: IUserRepository) {}
 
-    async execute({id}:DeleteUserDTO){
+  async execute({ id }: DeleteUserDTO) {
+    const userExist = await this.userRepository.findById(id);
 
-        const userExist = await this.userRepository.findById(id)
-
-        if(!userExist){
-            throw new ErrorUserDoesNotExist()
-        }
-
-        await this.userRepository.delete(id)
-
-        return
+    if (!userExist) {
+      throw new ErrorUserDoesNotExist();
     }
+
+    if (userExist.role == "Founder" && "Developer") {
+      throw new ErrorPermission();
+    }
+
+    await this.userRepository.delete(id);
+
+    return;
+  }
 }
